Add some description to major ticket attributes
