+++ package.json
@@ -22,20 +22,21 @@
   },
   "dependencies": {
     "@nestjs/cli": "^7.0.0",
-    "@nestjs/common": "^7.0.0",
+    "@nestjs/common": "^7.0.8",
     "@nestjs/config": "^0.4.0",
-    "@nestjs/core": "^7.0.0",
+    "@nestjs/core": "^7.0.8",
     "@nestjs/jwt": "^7.0.0",
     "@nestjs/mongoose": "^6.4.0",
     "@nestjs/passport": "^7.0.0",
-    "@nestjs/platform-express": "^7.0.0",
-    "@nestjs/swagger": "^4.5.1",
+    "@nestjs/platform-express": "^7.0.8",
+    "@nestjs/swagger": "^4.5.2",
     "@types/crypto-js": "^3.1.44",
+    "class-validator": "^0.12.1",
     "crypto-js": "^4.0.0",
     "dotenv": "^8.2.0",
     "env-cmd": "^10.1.0",
-    "mongodb": "^3.5.5",
-    "mongoose": "^5.9.7",
+    "mongodb": "^3.5.6",
+    "mongoose": "^5.9.9",
     "passport": "^0.4.1",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
@@ -46,28 +47,28 @@
   },
   "devDependencies": {
     "@nestjs/schematics": "^7.0.0",
-    "@nestjs/testing": "^7.0.0",
-    "@types/express": "^4.17.3",
+    "@nestjs/testing": "^7.0.8",
+    "@types/express": "^4.17.6",
     "@types/jest": "25.1.4",
-    "@types/mongoose": "^5.7.8",
-    "@types/node": "^13.9.1",
+    "@types/mongoose": "^5.7.12",
+    "@types/node": "^13.13.0",
     "@types/passport-jwt": "^3.0.3",
     "@types/passport-local": "^1.0.33",
     "@types/supertest": "^2.0.8",
-    "@typescript-eslint/eslint-plugin": "^2.23.0",
-    "@typescript-eslint/parser": "^2.23.0",
+    "@typescript-eslint/eslint-plugin": "^2.28.0",
+    "@typescript-eslint/parser": "^2.28.0",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-import": "^2.20.1",
     "husky": "^4.2.5",
-    "jest": "^25.1.0",
-    "lint-staged": "^10.1.3",
+    "jest": "^25.3.0",
+    "lint-staged": "^10.1.6",
     "mongodb-memory-server": "^6.5.2",
     "prettier": "^1.19.1",
     "supertest": "^4.0.2",
     "ts-jest": "25.2.1",
     "ts-loader": "^6.2.1",
-    "ts-node": "^8.6.2",
+    "ts-node": "^8.8.2",
     "tsconfig-paths": "^3.9.0",
     "typescript": "^3.7.4"
   },
