+++ src/ticketing/controller/tickets.controller.ts
@@ -1,51 +1,12 @@
 import { Body, Controller, Get, HttpCode, HttpException, HttpStatus, Logger, Param, Post } from '@nestjs/common';
-import { ApiTags } from '@nestjs/swagger';
+import { ApiTags, ApiProperty } from '@nestjs/swagger';
 import { HashingService } from '../../crypto/services/hashing.service';
 import { Address, Identity, Ticket, TicketsService, TicketStatus } from '../services/tickets.service';
-
-export class TicketRequestDto {
-    passportId: string;
-    reason: string;
-
-    startAddress: AddressDto;
-    endAddress: AddressDto;
-
-    validFromDateTime: Date;
-    validToDateTime: Date;
-}
-
-export class IdentityDto implements Identity {
-    hashedPassportId: string;
-}
-
-export class TicketResponseDto implements Ticket {
-    ticketId: string;
-
-    hashedPassportId: string;
-    reason: string;
-
-    startAddress: AddressDto;
-    endAddress: AddressDto;
-
-    validFromDateTime: Date;
-    validToDateTime: Date;
-
-    ticketStatus: TicketStatus;
-}
-
-export class AddressDto implements Address {
-    street: string;
-    houseNumber: string;
-    zipCode: string;
-    city: string;
-    country: string;
-}
+import { TicketRequestDto, TicketResponseDto, IdentityDto } from './tickets.dto';
 
 @ApiTags('ticket')
 @Controller('api/v1/tickets')
 export class TicketsController {
-    private readonly logger = new Logger(TicketsController.name);
-
     constructor(private readonly ticketsService: TicketsService, private readonly hashingService: HashingService) {}
 
     @Post()
